backend/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
django-revproxy==0.13.0
djangorestframework==3.17.1
idna==3.11
mysqlclient==2.2.8
requests==2.33.1
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
